+++ Venmo/src/test/java/com/braintreepayments/api/VenmoAccountUnitTest.java
@@ -16,13 +16,9 @@ public class VenmoAccountUnitTest {
     public void correctlyBuildsVenmoVaultRequest() throws JSONException {
         VenmoAccount sut = new VenmoAccount();
         sut.setNonce("some-nonce");
-        sut.setShouldValidate(true);
 
         JSONObject fullJson = sut.buildJSON();
         JSONObject venmoAccountJson = fullJson.getJSONObject("venmoAccount");
         assertEquals("some-nonce", venmoAccountJson.getString("nonce"));
-
-        JSONObject optionsJson = venmoAccountJson.getJSONObject("options");
-        assertTrue(optionsJson.getBoolean("validate"));
     }
 }