+++ CHANGELOG.md
@@ -37,7 +37,7 @@
   * `BraintreeClient` constructor no longer throws `InvalidArgumentException`
   * Make protected static member variables `OPTIONS_KEY`, `OPERATION_NAME_KEY` on `PaymentMethod` package-private
   * Make `PaymentMethod` constructor package-private
-  * Rename `setValidate` to `setShouldValidate` and move it from `PaymentMethod` base class to `Card`, `PayPalAccount` and `VenmoAccount` subclasses
+  * Rename `setValidate` to `setShouldValidate` and move it from `PaymentMethod` base class to `Card` subclass 
   * Make `buildJSON()` package-private for `PaymentMethod` base class and all subclasses
   * Remove `buildGraphQL()` method from `PaymentMethod` base class and all subclasses
   * Make `PaymentMethod` `Parcelable` constructor package-private